.kiro/specs/youtube-shorts-automation/requirements.txt
openai>=1.0.0
elevenlabs>=0.2.0
stabilityai>=0.1.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
